python-ai-backend/requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
opencv-python>=4.5.0
Flask-CORS>=3.0.0 
diffusers>=0.27.0
transformers>=4.38.0
rembg>=2.0.0
realesrgan>=0.3.0
torch>=2.2.0
pytesseract>=0.3.10
